rename the cpp file like pezy and add the makefile of this chapter!
