+++ ch17/ex17_10.cpp
