+++ ch17/ex17_14_15_16.cpp
