+++ ch17/ex17_17_18.cpp
