+++ ch17/ex17_19_20.cpp
