+++ ch17/ex17_1_2.cpp
