+++ ch17/ex17_21.cpp
