+++ ch17/ex17_28_29_30.cpp
