+++ ch17/ex17_9.cpp
@@ -22,7 +22,7 @@
 #include <iostream>
 #include <bitset>
 
-#include "Sales_data.h"
+//#include "Sales_data.h"
 
 
 int main()
